Libattachsql 2.0 rearranging
* asret renamed to attachsql_ret_t
* libattachsql-2.0 renamed to libattachsql2.  This is so the 2.1 release for
  example doesn't need a rename, semantic versioning means the include
  shouldn't break in major releases
* moved test cases to asql
