--- libattachsql-2.0/include.am
@@ -1,14 +1,0 @@
-# vim:ft=automake
-# included from Top Level Makefile.am
-# All paths should be given relative to the root
-
-nobase_include_HEADERS+= libattachsql-2.0/attachsql.h
-nobase_include_HEADERS+= libattachsql-2.0/connect.h
-nobase_include_HEADERS+= libattachsql-2.0/constants.h
-nobase_include_HEADERS+= libattachsql-2.0/error.h
-nobase_include_HEADERS+= libattachsql-2.0/pool.h
-nobase_include_HEADERS+= libattachsql-2.0/query.h
-nobase_include_HEADERS+= libattachsql-2.0/utility.h
-nobase_include_HEADERS+= libattachsql-2.0/statement.h
-nobase_include_HEADERS+= libattachsql-2.0/structs.h
-nobase_include_HEADERS+= libattachsql-2.0/visibility.h
