+++ libattachsql2/connect.h
