+++ libattachsql2/constants.h
