+++ libattachsql2/error.h
