+++ libattachsql2/pool.h
