+++ libattachsql2/structs.h
