+++ libattachsql2/utility.h
