+++ libattachsql2/visibility.h
