+++ src/buffer.h
@@ -50,7 +50,7 @@ struct buffer_st
 buffer_st *attachsql_buffer_create();
 void attachsql_buffer_free(buffer_st *buffer);
 size_t attachsql_buffer_get_available(buffer_st *buffer);
-asret_t attachsql_buffer_increase(buffer_st *buffer);
+attachsql_ret_t attachsql_buffer_increase(buffer_st *buffer);
 void attachsql_buffer_move_write_ptr(buffer_st *buffer, size_t len);
 size_t attachsql_buffer_unread_data(buffer_st *buffer);
 void attachsql_buffer_packet_read_end(buffer_st *buffer);
