+++ src/return.h
@@ -21,25 +21,25 @@
 extern "C" {
 #endif
 
-enum asret_t {
-  ASRET_OK,
-  ASRET_PARAMETER_ERROR,
-  ASRET_CONNECTING,
-  ASRET_DNS_ERROR,
-  ASRET_CONNECT_ERROR,
-  ASRET_OUT_OF_MEMORY_ERROR,
-  ASRET_NO_SCRAMBLE,
-  ASRET_USER_TOO_LONG,
-  ASRET_SCHEMA_TOO_LONG,
-  ASRET_NET_READ_ERROR,
-  ASRET_NET_SSL_ERROR,
-  ASRET_NET_WRITE_ERROR,
-  ASRET_PACKET_OUT_OF_SEQUENCE,
-  ASRET_BAD_PROTOCOL,
-  ASRET_NO_OLD_AUTH,
-  ASRET_BAD_SCRAMBLE,
-  ASRET_COMPRESSION_FAILURE,
-  ASRET_BAD_STMT_PARAMETER
+enum attachsql_ret_t {
+  ATTACHSQL_RET_OK,
+  ATTACHSQL_RET_PARAMETER_ERROR,
+  ATTACHSQL_RET_CONNECTING,
+  ATTACHSQL_RET_DNS_ERROR,
+  ATTACHSQL_RET_CONNECT_ERROR,
+  ATTACHSQL_RET_OUT_OF_MEMORY_ERROR,
+  ATTACHSQL_RET_NO_SCRAMBLE,
+  ATTACHSQL_RET_USER_TOO_LONG,
+  ATTACHSQL_RET_SCHEMA_TOO_LONG,
+  ATTACHSQL_RET_NET_READ_ERROR,
+  ATTACHSQL_RET_NET_SSL_ERROR,
+  ATTACHSQL_RET_NET_WRITE_ERROR,
+  ATTACHSQL_RET_PACKET_OUT_OF_SEQUENCE,
+  ATTACHSQL_RET_BAD_PROTOCOL,
+  ATTACHSQL_RET_NO_OLD_AUTH,
+  ATTACHSQL_RET_BAD_SCRAMBLE,
+  ATTACHSQL_RET_COMPRESSION_FAILURE,
+  ATTACHSQL_RET_BAD_STMT_PARAMETER
 };
 
 #ifdef __cplusplus
