+++ tests/connect_error.cc
@@ -17,7 +17,7 @@
 
 #include <yatl/lite.h>
 #include "version.h"
-#include <libattachsql-2.0/attachsql.h>
+#include <libattachsql2/attachsql.h>
 
 int main(int argc, char *argv[])
 {
@@ -25,37 +25,20 @@ int main(int argc, char *argv[])
   (void) argv;
   attachsql_connect_t *con;
   attachsql_error_t *error= NULL;
-  const char *data= "SHOW PROCESSLIST";
   attachsql_return_t aret= ATTACHSQL_RETURN_NONE;
-  attachsql_query_row_st *row;
-  uint16_t columns, col;
+  const char *data= "SHOW PROCESSLIST";
 
-  con= attachsql_connect_create("localhost", 3306, "test", "test", "", NULL);
-  attachsql_query_buffer_rows(con, true);
+  con= attachsql_connect_create("localhost", 3306, "bad_user", "test", "", NULL);
   attachsql_query(con, strlen(data), data, 0, NULL, &error);
-  while(aret != ATTACHSQL_RETURN_EOF)
+  while (aret != ATTACHSQL_RETURN_ERROR)
   {
     aret= attachsql_connect_poll(con, &error);
-    if (error && (attachsql_error_code(error) == 2002))
-    {
-      SKIP_IF_(true, "No MYSQL server");
-    }
-    else if (error)
-    {
-      ASSERT_FALSE_(true, "Error exists: %d", attachsql_error_code(error));
-    }
   }
-  columns= attachsql_query_column_count(con);
-  ASSERT_TRUE(attachsql_query_row_count(con) > 0);
-  while((row= attachsql_query_buffer_row_get(con)))
-  {
-    for (col=0; col < columns; col++)
-    {
-      printf("Column: %d, Length: %zu, Data: %.*s ", col, row[col].length, (int)row[col].length, row[col].data);
-    }
-    printf("\n");
-  }
-
+  SKIP_IF_((attachsql_error_code(error) == 2002), "Error not NULL");
+  ASSERT_EQ_(1045, attachsql_error_code(error), "Error code is wrong");
+  ASSERT_STREQL_("28000", attachsql_error_sqlstate(error), 5, "SQLSTATE is wrong");
+  ASSERT_STREQ_("Access denied for user 'bad_user'@'localhost' (using password: YES)", attachsql_error_message(error), "Message is wrong");
+  attachsql_error_free(error);
   attachsql_query_close(con);
   attachsql_connect_destroy(con);
 }
