+++ tests/multi_query.cc
@@ -17,7 +17,7 @@
 
 #include <yatl/lite.h>
 #include "version.h"
-#include <libattachsql-2.0/attachsql.h>
+#include <libattachsql2/attachsql.h>
 
 int main(int argc, char *argv[])
 {
@@ -25,16 +25,34 @@ int main(int argc, char *argv[])
   (void) argv;
   attachsql_connect_t *con;
   attachsql_error_t *error= NULL;
-  const char *data= "SHOW PROCESSLIST";
-  const char *data2= "SELECT ? as a, '?' as b, ? as c, ? as d, ? as e";
+  const char *data= "SHOW PROCESSLIST; SHOW PROCESSLIST";
   attachsql_return_t aret= ATTACHSQL_RETURN_NONE;
   attachsql_query_row_st *row;
-  attachsql_query_parameter_st param[5];
   uint16_t columns, col;
 
-  SKIP_IF_(!getenv("MYSQL_SOCK"), "The MYSQL_SOCK env variable must be set to run this test");
-  con= attachsql_connect_create(getenv("MYSQL_SOCK"), 0, "test", "test", "", NULL);
+  con= attachsql_connect_create("localhost", 3306, "test", "test", "", NULL);
+  attachsql_connect_set_option(con, ATTACHSQL_OPTION_MULTI_STATEMENTS, NULL);
   attachsql_query(con, strlen(data), data, 0, NULL, &error);
+  SKIP_IF_(error, "Error not NULL");
+  while(aret != ATTACHSQL_RETURN_EOF)
+  {
+    aret= attachsql_connect_poll(con, &error);
+    SKIP_IF_(error, "Multi-statement not supported");
+    if (aret == ATTACHSQL_RETURN_ROW_READY)
+    {
+      row= attachsql_query_row_get(con, &error);
+      columns= attachsql_query_column_count(con);
+      for (col=0; col < columns; col++)
+      {
+        printf("Column: %d, Length: %zu, Data: %.*s ", col, row[col].length, (int)row[col].length, row[col].data);
+      }
+      attachsql_query_row_next(con);
+      printf("\n");
+    }
+  }
+  attachsql_query_close(con);
+  aret= attachsql_query_next_result(con);
+  ASSERT_EQ_(ATTACHSQL_RETURN_PROCESSING, aret, "Unexpected next query status");
   while(aret != ATTACHSQL_RETURN_EOF)
   {
     aret= attachsql_connect_poll(con, &error);
@@ -49,53 +67,9 @@ int main(int argc, char *argv[])
       attachsql_query_row_next(con);
       printf("\n");
     }
-    if (error && (attachsql_error_code(error) == 2002))
-    {
-      SKIP_IF_(true, "No MYSQL server");
-    }
-    else if (error)
-    {
-      ASSERT_FALSE_(true, "Error exists: %d", attachsql_error_code(error));
-    }
   }
   attachsql_query_close(con);
-  const char *td= "test";
-  uint32_t tn= 45768;
-  float tf= 3.14159;
-  double tlf= 3.1415926;
-  param[0].type= ATTACHSQL_ESCAPE_TYPE_CHAR;
-  param[0].data= (char*)td;
-  param[0].length= strlen(td);
-  param[1].type= ATTACHSQL_ESCAPE_TYPE_CHAR_LIKE;
-  param[1].data= (char*)td;
-  param[1].length= strlen(td);
-  param[2].type= ATTACHSQL_ESCAPE_TYPE_INT;
-  param[2].data= &tn;
-  param[2].is_unsigned= true;
-  param[3].type= ATTACHSQL_ESCAPE_TYPE_FLOAT;
-  param[3].data= &tf;
-  param[4].type= ATTACHSQL_ESCAPE_TYPE_DOUBLE;
-  param[4].data= &tlf;
-  attachsql_query(con, strlen(data2), data2, 5, param, &error);
-  ASSERT_NULL_(error, "Error not NULL");
-  aret= ATTACHSQL_RETURN_NONE;
-  while(aret != ATTACHSQL_RETURN_EOF)
-  {
-    aret= attachsql_connect_poll(con, &error);
-    if (aret == ATTACHSQL_RETURN_ROW_READY)
-    {
-      row= attachsql_query_row_get(con, &error);
-      columns= attachsql_query_column_count(con);
-      ASSERT_EQ_(5, columns, "Column count unexpected");
-      ASSERT_STREQL_("test", row[0].data, 4, "Bad row data");
-      ASSERT_STREQL_("test", row[1].data, 4, "Bad row data");
-      ASSERT_STREQL_("45768", row[2].data, 5, "Bad row data");
-      ASSERT_STREQL_("3.14159", row[3].data, 7, "Bad float data");
-      ASSERT_STREQL_("3.1415926", row[4].data, 9, "Bad double data");
-      attachsql_query_row_next(con);
-      printf("\n");
-    }
-  }
-  attachsql_query_close(con);
+  attachsql_query_next_result(con);
+  ASSERT_EQ_(ATTACHSQL_RETURN_EOF, aret, "Unexpected next query status");
   attachsql_connect_destroy(con);
 }
