+++ tests/statement_twice.cc
@@ -17,7 +17,7 @@
 
 #include <yatl/lite.h>
 #include "version.h"
-#include <libattachsql-2.0/attachsql.h>
+#include <libattachsql2/attachsql.h>
 
 int main(int argc, char *argv[])
 {
@@ -27,7 +27,6 @@ int main(int argc, char *argv[])
   attachsql_error_t *error= NULL;
   const char *data= "SELECT ? as a, ? as b, CONVERT_TZ(FROM_UNIXTIME(1196440219),@@session.time_zone,'+00:00') as c";
   const char *data2= "hello world";
-  const char *data3= "dlrow olleh";
   attachsql_return_t aret= ATTACHSQL_RETURN_NONE;
   uint16_t columns;
 
@@ -76,8 +75,24 @@ int main(int argc, char *argv[])
       ASSERT_FALSE_(true, "Error exists: %d", attachsql_error_code(error));
     }
   }
-  attachsql_statement_set_string(con, 0, 11, data3, NULL);
-  attachsql_statement_set_int(con, 1, 654321, NULL);
+  attachsql_statement_close(con);
+  aret= ATTACHSQL_RETURN_NONE;
+  attachsql_statement_prepare(con, strlen(data), data, &error);
+  ASSERT_FALSE_(error, "Statement creation error");
+  while(aret != ATTACHSQL_RETURN_EOF)
+  {
+    aret= attachsql_connect_poll(con, &error);
+    if (error && (attachsql_error_code(error) == 2002))
+    {
+      SKIP_IF_(true, "No MYSQL server");
+    }
+    else if (error)
+    {
+      ASSERT_FALSE_(true, "Error exists: %d", attachsql_error_code(error));
+    }
+  }
+  attachsql_statement_set_double(con, 0, 3.14159, NULL);
+  attachsql_statement_set_datetime(con, 1, 2014, 10, 6, 21, 50, 20, 324560, NULL);
   attachsql_statement_execute(con, &error);
   aret= ATTACHSQL_RETURN_NONE;
   while(aret != ATTACHSQL_RETURN_EOF)
@@ -89,13 +104,12 @@ int main(int argc, char *argv[])
       attachsql_statement_row_get(con, &error);
       printf("Got %d columns\n", columns);
       size_t len;
-      char *col_data= attachsql_statement_get_char(con, 0, &len, &error);
+      char *col_data= attachsql_statement_get_char(con, 1, &len, &error);
       printf("Column 0: %.*s\n", (int)len, col_data);
-      printf("Column 1: %d\n", attachsql_statement_get_int(con, 1, &error));
-      ASSERT_EQ_(654321, attachsql_statement_get_int(con, 1, &error), "Column 1 result match fail");
-      ASSERT_STREQL_("dlrow olleh", col_data, len, "Column 0 result match fail");
-      col_data= attachsql_statement_get_char(con, 1, &len, &error);
-      ASSERT_STREQL_("654321", col_data, len, "Column 0 str conversion fail");
+      printf("Column 1: %f\n", attachsql_statement_get_double(con, 0, &error));
+      ASSERT_STREQL_("2014-10-06 21:50:20.324560", col_data, len, "Column 0 result match fail");
+      col_data= attachsql_statement_get_char(con, 0, &len, &error);
+      ASSERT_STREQL_("3.14159", col_data, 7, "Column 0 str conversion fail");
       col_data= attachsql_statement_get_char(con, 2, &len, &error);
       printf("Column 2: %.*s\n", (int)len, col_data);
       ASSERT_STREQL_("2007-11-30 16:30:19", col_data, len, "Column 2 str conversion fail");
@@ -107,5 +121,7 @@ int main(int argc, char *argv[])
     }
   }
   attachsql_statement_close(con);
+
+
   attachsql_connect_destroy(con);
 }
